Changes: - Update build number from `0` to `1`. - Convert [non win] from autotools/make to cmake/make. - Improve tests to ensure artifacts are in the right place. - Corrected license file and local contents.
Note: this is a version bump update to correct missing artifacts in
the package.

https://github.com/madler/zlib/tree/v1.2.12

https://github.com/madler/zlib/blob/v1.2.12/zlib.h

The license is `Zlib` and it's found in a header file.

https://github.com/madler/zlib/blob/v1.2.12/ChangeLog

Note, this is a build bump, not a version bump.

https://github.com/madler/zlib/issues

Note, this is a build bump, not a version bump. I didn't even look at the issues.

`ripgrep` is listed as required in the host section. I don't think
it's required, but I don't really have time for a full investigation
so I'm leaving it in.

Tests were updated to ensure artifacts exist in the package.

Shared libraries now included and confirmed with tests.